+++ tammx/types.h
@@ -17,12 +17,12 @@ struct BlockDimSpace;
 using BlockIndex = StrongNum<BlockDimSpace, int64_t>;
 using TensorRank = size_t;
 struct IrrepSpace;
-using Irrep = StrongNum<IrrepSpace, int>;
+using Irrep = StrongNum<IrrepSpace, int64_t>;
 struct SpinSpace;
-using Spin = StrongNum<SpinSpace, int>;
-using Sign = int;
+using Spin = StrongNum<SpinSpace, int64_t>;
+using Sign = int64_t;
 struct ProcSpace;
-using Proc = StrongNum<ProcSpace, int>;
+using Proc = StrongNum<ProcSpace, int64_t>;
 struct OffsetSpace;
 using Offset = StrongNum<OffsetSpace, int64_t>;
 using Size = Offset;
@@ -156,7 +156,7 @@ using TensorVec = BoundVec<T, maxrank>;
 
 using BlockDimVec = TensorVec<BlockIndex>;
 using DimTypeVec = TensorVec<DimType>;
-using PermVec = TensorVec<int>;
+using PermVec = TensorVec<int64_t>;
 
 } //namespace tammx
